# Conflicts:
#	test/Chirp.Web.Tests/UnitTests.cs

+++ test/Chirp.Web.Tests/IntegrationTests.cs
@@ -1,14 +1,7 @@
-﻿using System.ComponentModel;
-using System.Net;
-using Chirp.Core;
-using Chirp.Web;
+﻿using Chirp.Core;
 using Chirp.Infrastructure.Chirp.Repositories;
 using Chirp.Infrastructure.Chirp.Services;
-using Microsoft.AspNetCore.Mvc.Testing;
-using Microsoft.VisualStudio.TestPlatform.TestHost;
-using NUnit.Framework;
 using Xunit;
-using Xunit.Abstractions;
 using Assert = Xunit.Assert;
 
 namespace Chirp.Web.Tests;
@@ -17,14 +10,14 @@ namespace Chirp.Web.Tests;
 public class IntegrationTests : IAsyncLifetime
 {
 
-    private IAuthorRepository _authorrepo;
-    private ICheepRepository _cheeprepo;
-    private IFollowRepository _followrepo;
+    private IAuthorRepository? _authorrepo;
+    private ICheepRepository? _cheeprepo;
+    private IFollowRepository? _followrepo;
 
-    private TestUtilities _utils;
-    private CheepDbContext _context;
+    private TestUtilities? _utils;
+    private CheepDbContext? _context;
     
-    private IChirpService _service;
+    private IChirpService? _service;
     
     public async Task InitializeAsync()
     {
@@ -40,6 +33,11 @@ public class IntegrationTests : IAsyncLifetime
 
     public Task DisposeAsync()
     {
+        if (_utils == null)
+        {
+            return Task.CompletedTask;
+        }
+        _utils.CloseConnection();
         return Task.CompletedTask;
     }
     
@@ -47,6 +45,10 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task TestAddCheep()
     {
+        if (_utils == null || _cheeprepo == null || _authorrepo == null)
+        {
+            return;
+        }
         var cheepsBefore = await _cheeprepo.GetCheepsFromAuthor(0, "Mellie Yost");
         
         var author = await _authorrepo.GetAuthorByName("Mellie Yost");
@@ -67,8 +69,12 @@ public class IntegrationTests : IAsyncLifetime
 
     
     [Fact]
-    public async Task AddCheepchirpServiceNonExistingAuthor()
+    public async Task TestAddCheepchirpServiceNonExistingAuthor()
     {
+        if (_service == null || _utils == null || _authorrepo == null)
+        {
+            return;
+        }
         await _service.AddCheep("testest", "NewAuthor", "@newauthor.com");
     
         var author = await _authorrepo.GetAuthorByName("NewAuthor");
@@ -84,7 +90,11 @@ public class IntegrationTests : IAsyncLifetime
     {
         //We first make a user follow another user. Show that the cheeps should now be the sum of their cheeps in their private timeline
         //Octavio Wagganer(15) follow Mellie Yost(7)
-
+        if (_service == null)
+        {
+            return;
+        }
+        
         string authorname1 = "Octavio Wagganer";
         string authorname2 = "Mellie Yost";
 
@@ -99,6 +109,11 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task TestCheepHasFollowIfAuthorFollows()
     {
+        if (_service == null)
+        {
+            return;
+        }
+        
         string authorname1 = "Octavio Wagganer";
         string authorname2 = "Mellie Yost";
 
@@ -114,6 +129,11 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task TestDeleteFollows()
     {
+        if (_service == null || _authorrepo == null || _followrepo == null)
+        {
+            return;
+        }
+        
         await _authorrepo.CreateAuthor("erik", "hahaemail@gmail.com");
         await _authorrepo.CreateAuthor("lars", "bahaemail@gmail.com");
         await _followrepo.AddFollowing("erik", "lars");
@@ -131,6 +151,11 @@ public class IntegrationTests : IAsyncLifetime
     [Fact]
     public async Task TestCountingOfLikesOnCheep()
     {
+        if (_service == null)
+        {
+            return;
+        }
+        
         await _service.AddLike("Octavio Wagganer", 1);
         await _service.AddLike("Mellie Yost", 1);
 
@@ -141,8 +166,13 @@ public class IntegrationTests : IAsyncLifetime
 
 
     [Fact]
-    public async Task CanRemoveLike()
+    public async Task TestCanRemoveLike()
     {
+        if (_service == null)
+        {
+            return;
+        }
+        
         await _service.AddLike("Octavio Wagganer", 1);
         var likes1Amount = await _service.CountLikes(1);
         await _service.RemoveLike("Octavio Wagganer", 1);
@@ -153,8 +183,13 @@ public class IntegrationTests : IAsyncLifetime
 
 
     [Fact]
-    public async Task CanRemoveLikeData()
+    public async Task TestCanRemoveLikeData()
     {
+        if (_service == null)
+        {
+            return;
+        }
+        
         string author = "Octavio Wagganer";
         await _service.AddLike(author, 1);
         var likes1Amount = await _service.CountLikes(1);
@@ -164,5 +199,7 @@ public class IntegrationTests : IAsyncLifetime
         Assert.True(likes1Amount != likes2Amount);
         Assert.True(likes2Amount == 0);
     }
+    
+   
 
 }
